Move plugin to internal
