+++ internal/app/secretless/configurationmanagers/kubernetes/crd/manager.go
@@ -5,10 +5,10 @@ import (
 
 	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 	api_v1 "github.com/cyberark/secretless-broker/pkg/apis/secretless.io/v1"
 	"github.com/cyberark/secretless-broker/pkg/secretless/config"
 	config_v1 "github.com/cyberark/secretless-broker/pkg/secretless/config/v1"
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
 )
 
 type configurationManager struct {
