+++ internal/app/secretless/handlers/mysql/handler.go
@@ -2,13 +2,14 @@ package mysql
 
 import (
 	"fmt"
-	"github.com/cyberark/secretless-broker/internal/app/secretless/handlers/mysql/protocol"
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
 	"io"
 	"log"
 	"net"
 	"reflect"
 	"strconv"
+
+	"github.com/cyberark/secretless-broker/internal/app/secretless/handlers/mysql/protocol"
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 )
 
 // Handler connects a client to a backend. It uses the handler Config and Providers to
