+++ internal/app/secretless/handlers/pg/handler.go
@@ -7,7 +7,7 @@ import (
 	"net"
 
 	"github.com/cyberark/secretless-broker/internal/app/secretless/handlers/pg/protocol"
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 )
 
 // ClientOptions stores the option that were specified by the connection client.
