+++ internal/app/secretless/handlers/ssh/handler.go
@@ -9,7 +9,7 @@ import (
 
 	"golang.org/x/crypto/ssh"
 
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 )
 
 // ServerConfig is the configuration info for the target server
