+++ internal/app/secretless/listeners/http/listener.go
@@ -13,8 +13,8 @@ import (
 
 	"github.com/go-ozzo/ozzo-validation"
 
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 	config_v1 "github.com/cyberark/secretless-broker/pkg/secretless/config/v1"
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
 )
 
 // Listener listens for and handles new connections
