+++ internal/app/secretless/listeners/ssh/listener.go
@@ -14,9 +14,9 @@ import (
 	"github.com/go-ozzo/ozzo-validation"
 	"golang.org/x/crypto/ssh"
 
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 	"github.com/cyberark/secretless-broker/internal/pkg/util"
 	config_v1 "github.com/cyberark/secretless-broker/pkg/secretless/config/v1"
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
 )
 
 // Listener accepts SSH connections and MITMs them using a Handler.
