+++ internal/app/secretless/listeners/sshagent/listener.go
@@ -8,9 +8,9 @@ import (
 	"github.com/go-ozzo/ozzo-validation"
 	"golang.org/x/crypto/ssh/agent"
 
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 	"github.com/cyberark/secretless-broker/internal/pkg/util"
 	config_v1 "github.com/cyberark/secretless-broker/pkg/secretless/config/v1"
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
 )
 
 // Listener accepts ssh-agent connections and delegates them to the Handler.
