+++ internal/app/secretless/plugin/v1/configuration_manager.go
