+++ internal/app/secretless/plugin/v1/connection_manager.go
