+++ internal/app/secretless/plugin/v1/resolver.go
