+++ internal/app/secretless/plugins.go
@@ -4,8 +4,8 @@ import (
 	"github.com/cyberark/secretless-broker/internal/app/secretless/configurationmanagers"
 	"github.com/cyberark/secretless-broker/internal/app/secretless/handlers"
 	"github.com/cyberark/secretless-broker/internal/app/secretless/listeners"
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 	"github.com/cyberark/secretless-broker/internal/app/secretless/providers"
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
 )
 
 // InternalHandlers is the set of built-in handlers
