+++ internal/app/secretless/providers/file/provider.go
@@ -3,7 +3,7 @@ package file
 import (
 	"io/ioutil"
 
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 )
 
 // Provider reads the contents of the specified file.
