+++ internal/app/secretless/providers/providers.go
@@ -10,7 +10,7 @@ import (
 	literalProvider "github.com/cyberark/secretless-broker/internal/app/secretless/providers/literal"
 	vaultProvider "github.com/cyberark/secretless-broker/internal/app/secretless/providers/vault"
 
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 )
 
 // ProviderFactories contains the list of built-in provider factories
