+++ internal/app/secretless/providers/vault/provider.go
@@ -6,7 +6,7 @@ import (
 
 	vault "github.com/hashicorp/vault/api"
 
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 )
 
 // Provider provides data values from the Conjur vault.
