+++ internal/app/summon/command/command.go
@@ -12,7 +12,7 @@ import (
 	"github.com/cyberark/secretless-broker/internal/app/secretless"
 	"github.com/cyberark/secretless-broker/internal/pkg/plugin"
 
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 )
 
 // The code in this file operates at the CLI level; it reads CLI arguments and will exit the process.
