+++ internal/pkg/plugin/manager.go
@@ -20,9 +20,9 @@ import (
 	"gopkg.in/yaml.v2"
 
 	"github.com/cyberark/secretless-broker/internal/app/secretless"
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 	"github.com/cyberark/secretless-broker/pkg/secretless/config"
 	config_v1 "github.com/cyberark/secretless-broker/pkg/secretless/config/v1"
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
 )
 
 var _SupportedFileSuffixes = []string{".so"}
