+++ internal/pkg/plugin/resolver.go
@@ -6,8 +6,8 @@ import (
 	"strings"
 	"sync"
 
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 	config_v1 "github.com/cyberark/secretless-broker/pkg/secretless/config/v1"
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
 )
 
 // Resolver is used to instantiate providers and resolve credentials
