+++ internal/pkg/util/util.go
@@ -4,7 +4,7 @@ import (
 	"log"
 	"net"
 
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 )
 
 // Accept listeners for new connections from Listener `l` and notifies plugins
