+++ test/kubernetes_provider/kubernetes_provider_test.go
@@ -5,18 +5,18 @@ import (
 
 	. "github.com/smartystreets/goconvey/convey"
 
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 	"github.com/cyberark/secretless-broker/internal/app/secretless/providers"
 	"github.com/cyberark/secretless-broker/internal/app/secretless/providers/kubernetessecrets"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
-	testclient "k8s.io/client-go/kubernetes/fake"
 	"k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	testclient "k8s.io/client-go/kubernetes/fake"
 )
 
 func TestKubernetes_Provider(t *testing.T) {
 	var (
-		err error
-		provider plugin_v1.Provider
+		err                error
+		provider           plugin_v1.Provider
 		kubernetesProvider *kubernetessecrets.Provider
 	)
