+++ test/plugin/example/cmd/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 	"github.com/cyberark/secretless-broker/test/plugin/example"
 )
