+++ test/plugin/example/config_manager.go
@@ -1,8 +1,8 @@
 package example
 
 import (
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 	"github.com/cyberark/secretless-broker/pkg/secretless/config"
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
 )
 
 type configManager struct {
