+++ test/plugin/example/conn_manager.go
@@ -5,8 +5,8 @@ import (
 	"net"
 	"strconv"
 
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 	config_v1 "github.com/cyberark/secretless-broker/pkg/secretless/config/v1"
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
 )
 
 // connectionManager is an empty struct
