+++ test/plugin/example/handler.go
@@ -11,8 +11,8 @@ import (
 
 	"golang.org/x/crypto/ssh/agent"
 
+	plugin_v1 "github.com/cyberark/secretless-broker/internal/app/secretless/plugin/v1"
 	config_v1 "github.com/cyberark/secretless-broker/pkg/secretless/config/v1"
-	plugin_v1 "github.com/cyberark/secretless-broker/pkg/secretless/plugin/v1"
 )
 
 // connectionDetails stores the connection info to the real backend database.
